(PUP-6091) Activate interpolation for all Data Provider types
Before this commit, interpolation expressions in the form %{xx} was
only recognized by the Hiera Data Providers. This commit makes it
generally available to all Data Providers.
