+++ lib/puppet/pops.rb
@@ -27,7 +27,9 @@ module Puppet
     require 'puppet/pops/validation'
     require 'puppet/pops/issue_reporter'
     require 'puppet/pops/lookup'
+    require 'puppet/pops/lookup/interpolation'
     require 'puppet/pops/lookup/invocation'
+    require 'puppet/pops/lookup/sub_lookup'
     require 'puppet/pops/lookup/explainer'
 
     require 'puppet/pops/model/model'
