+++ spec/fixtures/unit/data_providers/environments/production/modules/abc/manifests/init.pp
@@ -2,8 +2,9 @@ class abc {
   include 'abc::def'
 }
 
-class abc::def ($test1, $test2, $test3) {
+class abc::def ($test1, $test2, $test3, $ipl ) {
   notify { $test1: }
   notify { $test2: }
   notify { $test3: }
+  notify { $ipl: }
 }
